requirements.txt
pdfkit
jinja2
openai
python-dotenv